requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==6.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
psycopg2==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.4
